Separate downloader interface
